requirements.txt
Flask==2.3.0
dnspython==2.4.2
gunicorn==20.1.0